Added Pinata::SCM::Git.local_modifications; Proper sandboxing

+++ spec/sandbox_repo_helper.rb
@@ -15,12 +15,17 @@ module SandboxRepoHelper
   end
 
   def in_sandbox(&block)
+    rspec_dir = FileUtils.pwd
     sandbox = File.join(Dir.tmpdir, 'pinata-test-sandbox')
-    FileUtils.rm_rf(sandbox)
+    FileUtils.rm_rf(sandbox) if Dir.exists? sandbox
     git = SandboxGit.new(Git.clone(SANDBOX_GITHUB, sandbox))
     git.checkout(SANDBOX_START_BRANCH)
     git.reset(SANDBOX_START_SHA)
-    FileUtils.cd(sandbox)
-    block.call(git)
+    begin
+      FileUtils.cd(sandbox)
+      block.call(git)
+    ensure
+      FileUtils.cd(rspec_dir)
+    end
   end
 end
